+++ functions/Backup-WTSettings.ps1
@@ -14,32 +14,32 @@ Function Backup-WTSetting {
         [switch]$Passthru
     )
 
-    Write-Verbose "Starting $($myinvocation.MyCommand)"
+    Write-Verbose "[$((Get-Date).TimeofDay)] Starting $($myinvocation.MyCommand)"
     $json = "$ENV:Userprofile\AppData\Local\Packages\Microsoft.WindowsTerminal_8wekyb3d8bbwe\LocalState\settings.json"
 
     if (Test-Path $json) {
 
-        Write-Verbose "Backing up $json to $Destination"
-        Write-Verbose "Get existing backups from $Destination and saving as an array sorted by name"
+        Write-Verbose "[$((Get-Date).TimeofDay)] Backing up $json to $Destination"
+        Write-Verbose "[$((Get-Date).TimeofDay)] Get existing backups from $Destination and saving as an array sorted by name"
 
         [array]$bak = Get-ChildItem -path $Destination\settings.bak*.json | Sort-Object -Property LastWriteTimeString
 
         if ($bak.count -eq 0) {
-            Write-Verbose "Creating first backup copy."
+            Write-Verbose "[$((Get-Date).TimeofDay)] Creating first backup copy."
             [int]$new = 1
         }
         else {
             $bak | Out-String | Write-Verbose
             #get the numeric value
             [int]$counter = ([regex]"\d+").match($bak[-1]).value
-            Write-Verbose "Last backup is #$counter"
+            Write-Verbose "[$((Get-Date).TimeofDay)] Last backup is #$counter"
 
             [int]$new = $counter + 1
-            Write-Verbose "Creating backup copy $new"
+            Write-Verbose "[$((Get-Date).TimeofDay)] Creating backup copy $new"
         }
 
         $backup = Join-Path -path $Destination -ChildPath "settings.bak$new.json"
-        Write-Verbose "Creating backup $backup"
+        Write-Verbose "[$((Get-Date).TimeofDay)] Creating backup $backup"
         Copy-Item -Path $json -Destination $backup
 
         #update the list of backups sorted by age and delete extras
@@ -50,7 +50,7 @@ Function Backup-WTSetting {
         Select-Object -Skip $Limit | Remove-Item
 
         #renumber backup files
-        Write-Verbose "Renumbering backup files"
+        Write-Verbose "[$((Get-Date).TimeofDay)] Renumbering backup files"
         <#
     You can't rename a file if it will conflict with an existing file so files will be copied
     to a temp folder with a new name, the old file deleted and then the copy restored
@@ -62,14 +62,14 @@ Function Backup-WTSetting {
             $n++
             $temp = Join-Path -path $env:TEMP -ChildPath "settings.bak$n.json"
 
-            Write-Verbose "Copying temp file to $temp"
+            Write-Verbose "[$((Get-Date).TimeofDay)] Copying temp file to $temp"
             $_ | Copy-Item -Destination $temp
 
-            Write-Verbose "Removing $($_.name)"
+            Write-Verbose "[$((Get-Date).TimeofDay)] Removing $($_.name)"
             $_ | Remove-Item
 
         } -end {
-            Write-Verbose "Restoring temp files to $Destination"
+            Write-Verbose "[$((Get-Date).TimeofDay)] Restoring temp files to $Destination"
             Get-ChildItem -Path "$env:TEMP\settings.bak*.json" | Move-Item -Destination $Destination -PassThru:$passthru
         }
 
@@ -77,5 +77,5 @@ Function Backup-WTSetting {
     else {
         Write-Warning "Failed to find expected settings file $json"
     }
-        Write-Verbose "Ending $($myinvocation.MyCommand)"
+        Write-Verbose "[$((Get-Date).TimeofDay)] Ending $($myinvocation.MyCommand)"
 }