+++ functions/Get-WTKeybindings.ps1
@@ -8,29 +8,30 @@ Function Get-WTKeyBinding {
         [string]$Format = "None"
     )
 
+    Write-Verbose "[$((Get-Date).TimeofDay)] Starting $($MyInvocation.MyCommand)"
     #use a list object to make it easier to remove duplicate keybindings
     $list = [System.Collections.Generic.List[Object]]::new()
 
-    Write-Verbose "Getting WindowsTerminal Appx package"
+    Write-Verbose "[$((Get-Date).TimeofDay)] Getting WindowsTerminal Appx package"
     $install =(Get-AppxPackage -name Microsoft.WindowsTerminal).InstallLocation
-    Write-Verbose "Getting defaults.json file"
+    Write-Verbose "[$((Get-Date).TimeofDay)] Getting defaults.json file"
     $defaults = Join-Path -path $install -ChildPath defaults.json
 
-    Write-Verbose "Getting default Windows Terminal settings from $defaults"
+    Write-Verbose "[$((Get-Date).TimeofDay)] Getting default Windows Terminal settings from $defaults"
     #strip out the // comments since they aren't valid json
     $defaultsettings = Get-Content -path $defaults | Where-Object {$_ -notmatch "//"} | ConvertFrom-Json
 
     #get the keybindings and add a property that indicates where the setting came from.
-    Write-Verbose "Parsing default keybindings"
+    Write-Verbose "[$((Get-Date).TimeofDay)] Parsing default keybindings"
     $keys = $defaultsettings.keybindings |
     parsesetting |
     Select-Object -Property *, @{Name = "Source"; Expression = {"Defaults"}}
 
-    Write-Verbose "Found $($keys.count) default keybindings"
+    Write-Verbose "[$((Get-Date).TimeofDay)] Found $($keys.count) default keybindings"
     #add the keybinding objects to the list
     $list.AddRange($keys)
 
-    Write-Verbose "Getting user settings"
+    Write-Verbose "[$((Get-Date).TimeofDay)] Getting user settings"
     $settingsjson = "$ENV:Userprofile\AppData\Local\Packages\Microsoft.WindowsTerminal_8wekyb3d8bbwe\LocalState\settings.json"
     if (Test-Path $settingsjson) {
         $settings = Get-Content -path $settingsjson | Where-Object {$_ -notmatch "//"} | ConvertFrom-Json
@@ -40,13 +41,13 @@ Function Get-WTKeyBinding {
         $user = $settings.keybindings |
         parsesetting |
         Select-Object -Property *, @{Name = "Source"; Expression = {"Settings"}}
-        Write-Verbose "Found $($keys.count) user keybindings"
+        Write-Verbose "[$((Get-Date).TimeofDay)] Found $($keys.count) user keybindings"
 
         #if there is a duplicate key binding, remove the default
         foreach ($item in $user) {
             $existing = $list.where( {$_.keys -eq $item.keys})
             if ($existing) {
-                Write-Verbose "Detected an override of $($existing| Out-String)"
+                    Write-Verbose "[$((Get-Date).TimeofDay)] Detected an override of $($existing| Out-String)"
                 [void]($list.Remove($existing))
             }
             #add the entry
@@ -55,6 +56,7 @@ Function Get-WTKeyBinding {
     } #if keybindings
     }
 
+    Write-Verbose "[$((Get-Date).TimeofDay)] Formatting keybinding settings as $Format"
     Switch ($Format) {
         "Table" {
             $list | Sort-Object -Property Source, Keys |
@@ -71,4 +73,5 @@ Function Get-WTKeyBinding {
             $list
         }
     }
+    Write-Verbose "[$((Get-Date).TimeofDay)] Ending $($MyInvocation.MyCommand)"
 }