+++ functions/open-wtdefaults.ps1
@@ -4,11 +4,14 @@ Function Open-WTDefault {
     [CmdletBinding()]
     Param()
 
+    Write-Verbose "[$((Get-Date).TimeofDay)] Starting $($MyInvocation.MyCommand)"
+    Write-Verbose "[$((Get-Date).TimeofDay)] Getting AppxPackage for Microsoft.WindowsTerminal"
     $app = Get-AppxPackage Microsoft.WindowsTerminal
 
     if (Test-Path $app.InstallLocation) {
         $json = Join-Path -path $app.installlocation -ChildPath defaults.json
         if (Test-Path $json) {
+            Write-Verbose "[$((Get-Date).TimeofDay)] Opening $json"
             Invoke-Item $json
         }
         else {
@@ -16,6 +19,8 @@ Function Open-WTDefault {
         }
     }
     else {
-        Throw "WindowsTerminal is not installed"
+        Throw "Windows Terminal is not installed."
     }
+
+    Write-Verbose "[$((Get-Date).TimeofDay)] Ending $($MyInvocation.MyCommand)"
 }